requirement.txt
flask
python-dotenv
pymongo
passlib
wtforms
datetime
Flask-SQLAlchemy
Flask-Login
psycopg2
Werkzeug
